Disabled main page removing

+++ app/controllers/caffeine/admin/pages_controller.rb
@@ -34,6 +34,8 @@ module Caffeine
       end
 
       def destroy
+        return unless page.main?
+
         page.destroy
         respond_with :admin, page
       end
